Made patient app-header consistent with  medic home
When the patient clicks on the 'bird' icon, it reroutes them to the patient home route. Currently, all the patient UI is on the patient home route, but this onclick has been implemented in case we need to add things in the future. Also, added spacing at the end of the data analysis reports.