Add Set Diagnosis Vue initial attempt
Setting diagnosis has been Added. Currently, the vue will query the DB for the conditions that are within our scope. 

Todo: Add animation to show diagnosis saving. Sytling with Bulma. Need to do a validation check when the user inputs a diagnosis to ensure it is a diagnosis CareAway supports. Otherwise, print a message. Communicate with Taylor to ask where this will go as it is currently a modal on the medic-home.