Refactored patient selector in DA reports to its own component
The patient selector can now be used to select a patient by the MP. This component is to be used to select patients on the data analysis reports view and the treatment creator view. After a patient is selected, the component will emit 'selected' which will be received and handled by the parent component.